Revissio general

+++ v2/Source/Controllers/Display/Drivers/ILI9341/eosDevice_ILI9341.cpp
@@ -21,7 +21,7 @@ Device_ILI9341::Device_ILI9341() {
 ///
 void Device_ILI9341::writeScript(
     const uint8_t *script,
-    uint16_t scriptSize) {
+    unsigned scriptSize) {
 
     eosAssert(script != nullptr);
     eosAssert(scriptSize > 0);
